fix: case where request is made for position on chain angle doesnt support
