+++ apps/evm/src/ui/pool/PositionView.tsx
@@ -47,8 +47,8 @@ import { Chain } from 'sushi/chain'
 import { Amount } from 'sushi/currency'
 import { formatUSD } from 'sushi/format'
 
+import { isAngleEnabledChainId } from 'sushi/config'
 import { getAddress } from 'viem'
-import { isAngleEnabledChainId } from '../../config'
 import { Bound } from '../../lib/constants'
 import {
   formatTickPrice,
