+++ packages/sushi/src/config/index.ts
@@ -1,3 +1,4 @@
+export * from './angle'
 export * from './bentobox'
 export * from './extractor'
 export * from './route-processor'
